refactor: extract event creation rules to event creator module (#16194)
