+++ platform-sdk/event-creator-impl/src/main/java/org/hiero/event/creator/impl/EventCreationConfig.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.swirlds.platform.event.creation;
+package org.hiero.event.creator.impl;
 
 import com.swirlds.config.api.ConfigData;
 import com.swirlds.config.api.ConfigProperty;
