+++ platform-sdk/event-creator-impl/src/main/java/org/hiero/event/creator/impl/rules/PlatformHealthRule.java
@@ -14,16 +14,17 @@
  * limitations under the License.
  */
 
-package com.swirlds.platform.event.creation.rules;
+package org.hiero.event.creator.impl.rules;
 
 import static com.swirlds.common.utility.CompareTo.isLessThanOrEqualTo;
-import static com.swirlds.platform.event.creation.EventCreationStatus.OVERLOADED;
+import static org.hiero.event.creator.EventCreationStatus.OVERLOADED;
 
-import com.swirlds.platform.event.creation.EventCreationStatus;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import java.time.Duration;
 import java.util.Objects;
 import java.util.function.Supplier;
+import org.hiero.event.creator.EventCreationRule;
+import org.hiero.event.creator.EventCreationStatus;
 
 /**
  * A rule that permits event creation only when the platform is considered to be in a healthy state.
