+++ platform-sdk/event-creator/src/main/java/module-info.java
@@ -1,3 +1,5 @@
 module org.hiero.event.creator {
     exports org.hiero.event.creator;
+
+    requires static transitive com.github.spotbugs.annotations;
 }
