+++ platform-sdk/event-creator/src/main/java/org/hiero/event/creator/EventCreationRule.java
@@ -14,9 +14,8 @@
  * limitations under the License.
  */
 
-package com.swirlds.platform.event.creation.rules;
+package org.hiero.event.creator;
 
-import com.swirlds.platform.event.creation.EventCreationStatus;
 import edu.umd.cs.findbugs.annotations.NonNull;
 
 /**
