+++ platform-sdk/swirlds-platform-core/src/test/java/com/swirlds/platform/config/PlatformConfigurationExtensionTests.java
@@ -26,7 +26,7 @@ class PlatformConfigurationExtensionTests {
     @Test
     void testIfAllConfigDataTypesAreRegistered() {
         // given
-        final var allRecordsFound = ConfigUtils.loadAllConfigDataRecords(Set.of("com.swirlds"));
+        final var allRecordsFound = ConfigUtils.loadAllConfigDataRecords(Set.of("com.swirlds", "org.hiero"));
         final var extension = new PlatformConfigurationExtension();
 
         // when
