+++ platform-sdk/swirlds-unit-tests/core/swirlds-platform-test/build.gradle.kts
@@ -33,6 +33,8 @@ testModuleInfo {
     requires("org.junit.jupiter.params")
     requires("org.mockito.junit.jupiter")
     requires("com.swirlds.metrics.api")
+    requires("org.hiero.event.creator")
+    requires("org.hiero.event.creator.impl")
     requires("org.mockito")
     requires("org.hiero.consensus.gossip")
     requiresStatic("com.github.spotbugs.annotations")
